mid-measure methods moved to QubitManaged
