+++ pennylane_lightning/core/src/simulators/lightning_qubit/StateVectorLQubit.hpp
@@ -22,6 +22,7 @@
 
 #pragma once
 #include <complex>
+#include <random>
 #include <unordered_map>
 
 #include "CPUMemoryModel.hpp"
@@ -36,6 +37,7 @@ namespace {
 using Pennylane::LightningQubit::Util::Threading;
 using Pennylane::Util::CPUMemoryModel;
 using Pennylane::Util::exp2;
+using Pennylane::Util::squaredNorm;
 using namespace Pennylane::LightningQubit::Gates;
 } // namespace
 /// @endcond
@@ -74,6 +76,8 @@ class StateVectorLQubit : public StateVectorBase<PrecisionT, Derived> {
         std::unordered_map<ControlledGeneratorOperation, KernelType>;
     using ControlledMatrixKernelMap =
         std::unordered_map<ControlledMatrixOperation, KernelType>;
+
+    std::mt19937_64 rng;
 
     GateKernelMap kernel_for_gates_;
     GeneratorKernelMap kernel_for_generators_;
@@ -641,7 +645,106 @@ class StateVectorLQubit : public StateVectorBase<PrecisionT, Derived> {
     inline auto applyMidMeasureMP(const std::vector<size_t> &wires) -> int {
         PL_ABORT_IF_NOT(wires.size() == 1,
                         "MidMeasureMP should have a single wire.")
-        return static_cast<Derived *>(this)->measure(wires[0]);
+        return measure(wires[0]);
+    }
+
+    /**
+     * @brief Set the seed of the internal random generator
+     *
+     * @param seed Seed
+     */
+    void seed(const size_t seed) { rng.seed(seed); }
+
+    /**
+     * @brief Sample 0 or 1 for given probabilities
+     *
+     * @param probs Probabilities of 0.
+     */
+    const auto random_sample(const PrecisionT prob_0) {
+        std::discrete_distribution<int> distrib({prob_0, 1. - prob_0});
+        return distrib(rng);
+    }
+
+    /**
+     * @brief Calculate probabilities of measuring 0 or 1 on a specific wire
+     *
+     * @param wire Wire
+     * @return std::vector<PrecisionT> Both probabilities of getting 0 or 1
+     */
+    auto probs(const std::size_t wire) -> std::vector<PrecisionT> {
+        auto *arr = this->getData();
+        const size_t stride = pow(2, this->num_qubits_ - (1 + wire));
+        const size_t vec_size = pow(2, this->num_qubits_);
+        const auto section_size = vec_size / stride;
+        const auto half_section_size = section_size / 2;
+
+        PrecisionT prob_0{0.};
+        // zero half the entries
+        // the "half" entries depend on the stride
+        // *_*_*_*_ for stride 1
+        // **__**__ for stride 2
+        // ****____ for stride 4
+        const size_t k = 0;
+        for (size_t idx = 0; idx < half_section_size; idx++) {
+            for (size_t ids = 0; ids < stride; ids++) {
+                auto v = stride * (k + 2 * idx) + ids;
+                prob_0 += std::norm(arr[v]);
+            }
+        }
+
+        const std::vector<PrecisionT> probs{prob_0, PrecisionT(1.) - prob_0};
+        return probs;
+    }
+
+    /**
+     * @brief Collapse the state vector as after having measured one of the
+     * qubits
+     *
+     * @param wire Wire to collapse.
+     * @param branch Branch 0 or 1.
+     */
+    void collapse(const std::size_t wire, const bool branch) {
+        auto *arr = this->getData();
+        const size_t stride = pow(2, this->num_qubits_ - (1 + wire));
+        const size_t vec_size = pow(2, this->num_qubits_);
+        const auto section_size = vec_size / stride;
+        const auto half_section_size = section_size / 2;
+
+        // zero half the entries
+        // the "half" entries depend on the stride
+        // *_*_*_*_ for stride 1
+        // **__**__ for stride 2
+        // ****____ for stride 4
+        const size_t k = branch ? 0 : 1;
+        for (size_t idx = 0; idx < half_section_size; idx++) {
+            for (size_t ids = 0; ids < stride; ids++) {
+                auto v = stride * (k + 2 * idx) + ids;
+                arr[v] = {0., 0.};
+            }
+        }
+
+        PrecisionT norm = std::sqrt(squaredNorm(arr, this->getLength()));
+
+        if (norm > std::numeric_limits<PrecisionT>::epsilon() * 1e2) {
+            std::complex<PrecisionT> inv_norm = 1. / norm;
+            for (size_t k = 0; k < this->getLength(); k++) {
+                arr[k] *= inv_norm;
+            }
+        }
+    }
+
+    /**
+     * @brief Measure one of the qubits and collapse the state accordingly
+     *
+     * @param wire Wire to measure.
+     * @param branch Branch 0 or 1.
+     */
+    const auto measure(const std::size_t wire) {
+
+        std::vector<PrecisionT> probs_ = probs(wire);
+        auto sample = random_sample(probs_[0]);
+        collapse(wire, sample);
+        return sample;
     }
 };
 } // namespace Pennylane::LightningQubit
